askapi/Backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Limiter==3.5.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.93.1
anthropic==0.64.0
ordered-set==4.1.0
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
redis==5.2.1
requests==2.32.4
rich==13.9.4
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
supabase==2.8.0
PyYAML>=6.0
PyJWT==2.8.0
python-dateutil==2.9.0
google-cloud-vision==3.4.4
google-auth==2.23.4
Pillow==10.4.0
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
python-pptx==0.6.23

